+++ public_modules/home.html
@@ -14,7 +14,7 @@
     /> -->
     <script src="https://cdn.jsdelivr.net/npm/js-cookie@3.0.5/dist/js.cookie.min.js"></script>
     <script src="./script.mjs" type="module"></script>
-    <script src="./modules/getSloikId.js"></script>
+    <script src="./modules/cookies/getSloikId.js"></script>
   </head>
   <body>
     <div id="popupContainer">
