+++ public_modules/modules/home/sloikGeneration.mjs
@@ -1,5 +1,5 @@
-import { showPopup, hidePopup, showBulkEditPopup } from "./popup.mjs";
-import { getCookie, setCookie, setupSloikCookies } from "./cookies.mjs";
+import { hidePopup, showBulkEditPopup } from "./popup.mjs";
+import { getCookie, setCookie, setupSloikCookies } from "./cookies/cookies.mjs";
 import { deleteSloik, resetButtons, tryToDeleteSloik } from "./homePage.mjs";
 import { editBtnSvg } from "../const/editSvg.mjs";
 import { numberRegExp, stringRegExp } from "../const/regExp.mjs";
